Remove duplicated specs
* There is already library/socket/constants/constants_spec.rb
