+++ library/socket/constants/constants_spec.rb
@@ -86,4 +86,17 @@ describe "Socket::Constants" do
       Socket::Constants.should have_constant(c)
     end
   end
+
+  platform_is_not :windows do
+    it 'defines SCM options' do
+      Socket::Constants.should have_constant('SCM_CREDENTIALS')
+    end
+
+    it 'defines error options' do
+      consts = ["EAI_ADDRFAMILY", "EAI_NODATA"]
+      consts.each do |c|
+        Socket::Constants.should have_constant(c)
+      end
+    end
+  end
 end
